refactor(rust-bridge): drive both Python-embedded semantic caches through one execution
Redis-semantic and Valkey-semantic operations now share one SemanticExecution
body: await the Python embedder, seed the task-local vector, run the native
backend, repeat per batch entry. Valkey drops its with_embedder path in favor
of the same seeded embedder, and each backend keeps its own embedding-failure
policy. PythonEmbedder exposes one call shape. Redis-semantic thresholds are
compared at the backend's f32 width, which un-breaks the redis-stack parity
tests that a 0.8 facade threshold failed before this branch.

+++ litellm-rust/crates/python-bridge/src/cache/config.rs
@@ -338,7 +338,7 @@ impl CacheBackendConfig {
             },
             Self::RedisSemantic(config) => BackendIdentity::RedisSemantic {
                 index_name: config.index_name.clone(),
-                similarity_threshold: config.similarity_threshold,
+                similarity_threshold: config.similarity_threshold as f32,
             },
             Self::QdrantSemantic(config) => BackendIdentity::QdrantSemantic {
                 collection_name: config.collection_name.clone(),
